refactor: rename command `use` to `switch`
