+++ src/commands/switch.rs
@@ -1,6 +1,4 @@
 use crate::utils::deno_bin_path;
-use crate::utils::is_china_mainland;
-use crate::version::dvmrc_version;
 use anyhow::Result;
 use semver_parser::version::{parse as semver_parse, Version};
 use std::env;
@@ -9,36 +7,13 @@ use std::path::{Path, PathBuf};
 use std::process::Command;
 use which::which;
 
-pub fn exec(version: Option<String>) -> Result<()> {
-  fn get_latest_version() -> Result<Version> {
-    println!("Checking for latest version");
-    let response = if is_china_mainland() {
-      tinyget::get("https://dl.deno.js.cn/release-latest.txt").send()?
-    } else {
-      tinyget::get("https://dl.deno.land/release-latest.txt").send()?
-    };
-
-    let body = response.as_str()?;
-    let v = body.trim().replace('v', "");
-    println!("The latest version is v{}", &v);
-    Ok(semver_parse(&v).unwrap())
-  }
-
-  let version = version.unwrap_or_else(|| {
-    println!("No version input detect, try to use version in .dvmrc file");
-    dvmrc_version().unwrap_or_else(|| {
-      println!("No version in .dvmrc file, try to use latest version");
-      get_latest_version().unwrap().to_string()
-    })
-  });
-
-  let used_version = match semver_parse(&version) {
-    Ok(ver) => ver,
-    Err(_) => {
-      eprintln!("Invalid semver");
-      std::process::exit(1)
-    }
-  };
+pub fn exec(version: String) -> Result<()> {
+  let used_version = semver_parse(version.as_str()).map_err(|_| {
+    anyhow::Error::msg(format!(
+      "The version you input `{}` is not a valid semver version",
+      &version
+    ))
+  })?;
 
   let new_exe_path = deno_bin_path(&used_version);
