+++ src/commands/use.rs
@@ -1,0 +1,6 @@
+use anyhow::Result;
+
+pub fn exec(_version: String) -> Result<()> {
+  eprintln!("This command is not implemented yet");
+  std::process::exit(1);
+}
